Criação do middleware para tratar cep e outros tratamentos de erros

+++ src/services/storeService.js
@@ -15,7 +15,7 @@ const cepToCoordenates_js_1 = require("../utils/cepToCoordenates.js");
 const cepData_js_1 = require("../utils/cepData.js");
 const distanceCalculate_js_1 = require("../utils/distanceCalculate.js");
 const loggers_js_1 = require("../utils/loggers.js");
-//Criação da loja com dados manuais --- não finalizado(doing)
+//Criação da loja com dados manuais
 const createStore = (data) => __awaiter(void 0, void 0, void 0, function* () {
     try {
         const store = new storeSchema_js_1.Store();
@@ -24,16 +24,16 @@ const createStore = (data) => __awaiter(void 0, void 0, void 0, function* () {
         store.numero = data.numero;
         const adress = `${data.endereco.logradouro}, ${data.endereco.cidade}, ${data.endereco.estado}, Brasil`;
         store.coordenadas = yield (0, cepToCoordenates_js_1.getCoordenates)(adress);
-        store.save();
+        yield store.save();
         loggers_js_1.logger.info("Loja criada: ", store.toJSON);
         return store;
     }
     catch (error) {
         loggers_js_1.logger.error("Não foi possivel criar a loja: ", error);
-        throw new Error('Someting goes wrong!');
+        throw new Error('Não foi possivel criar a loja!');
     }
 });
-//Criação da loja com o CEP dado --- não finalizado(doing)
+//Criação da loja com o CEP dado
 const createStoreByCep = (data, cep) => __awaiter(void 0, void 0, void 0, function* () {
     try {
         const newData = yield (0, cepData_js_1.cepInfos)(cep.toString());
@@ -46,16 +46,16 @@ const createStoreByCep = (data, cep) => __awaiter(void 0, void 0, void 0, functi
         store.endereco.logradouro = newData.logradouro;
         const adress = `${newData.logradouro}, ${newData.localidade}, ${newData.estado}, Brasil`;
         store.coordenadas = yield (0, cepToCoordenates_js_1.getCoordenates)(adress);
-        store.save();
-        loggers_js_1.logger.info("Loja criada: ", store.toJSON);
+        yield store.save();
+        loggers_js_1.logger.info("Loja criada: ", store.toJSON());
         return store;
     }
     catch (error) {
         loggers_js_1.logger.error("Não foi possivel criar a loja: ", error);
-        throw new Error('Someting goes wrong!');
+        throw new Error('Não foi possivel criar a loja!');
     }
 });
-//Pegando todas as lojas do banco de dados --- não finalizado(doing)
+//Pegando todas as lojas do banco de dados 
 const getAllStores = () => __awaiter(void 0, void 0, void 0, function* () {
     try {
         const stores = yield storeSchema_js_1.Store.find();
@@ -63,45 +63,58 @@ const getAllStores = () => __awaiter(void 0, void 0, void 0, function* () {
     }
     catch (error) {
         loggers_js_1.logger.error("Não foi possivel acessar todas as lojas: ", error);
-        throw new Error('Someting goes wrong!');
+        throw new Error('Não foi possivel encontrar as lojas!');
     }
 });
-//Pegando uma a loja do banco de dados pelo ID --- não finalizado(doing)
+//Pegando uma a loja do banco de dados pelo ID 
 const getStoreById = (id) => __awaiter(void 0, void 0, void 0, function* () {
     try {
         const store = yield storeSchema_js_1.Store.findById(id);
+        if (!store) {
+            throw new Error('Loja não encontrada!');
+        }
         return store;
     }
     catch (error) {
         loggers_js_1.logger.error("Não foi possivel acessar a loja: ", error);
-        throw new Error('Someting goes wrong!');
+        throw new Error('Não foi possivel encontrar a loja.');
     }
 });
-//Atualizando uma a loja do banco de dados pelo ID --- não finalizado(doing)
-const updateStore = (id, params) => __awaiter(void 0, void 0, void 0, function* () {
+//Atualizando uma a loja do banco de dados pelo ID
+const updateStore = (id, body) => __awaiter(void 0, void 0, void 0, function* () {
     try {
-        const updateStore = yield storeSchema_js_1.Store.findByIdAndUpdate(id, params, {
+        const store = storeSchema_js_1.Store.findById(id);
+        if (!store) {
+            throw new Error('Loja não encontrada!');
+        }
+        const updateStore = yield storeSchema_js_1.Store.findByIdAndUpdate(id, body, {
             new: true,
             runValidators: true
         });
+        const adress = `${updateStore.endereco.logradouro}, ${updateStore.endereco.cidade}, ${updateStore.endereco.estado}, Brasil`;
+        updateStore.coordenadas = yield (0, cepToCoordenates_js_1.getCoordenates)(adress);
         return updateStore;
     }
     catch (error) {
         loggers_js_1.logger.error("Não foi possivel editar a loja: ", error);
-        throw new Error('Someting goes wrong!');
+        throw new Error('Erro ao editar a loja!');
     }
 });
-//Deletando uma a loja do banco de dados pelo ID --- não finalizado(doing)
+//Deletando uma a loja do banco de dados pelo ID
 const deleteStore = (id) => __awaiter(void 0, void 0, void 0, function* () {
     try {
+        const store = storeSchema_js_1.Store.findById(id);
+        if (!store) {
+            throw new Error('Loja não encontrada!');
+        }
         yield storeSchema_js_1.Store.findByIdAndRemove(id);
     }
     catch (error) {
         loggers_js_1.logger.error("Não foi possivel deletar a loja: ", error);
-        throw new Error('Someting goes wrong!');
+        throw new Error('Erro ao deletar a loja!');
     }
 });
-//Query de lojas a 100 km do cep dado pelo usuario --- não finalizado(doing)
+//Query de lojas a 100 km do cep dado pelo usuario
 const findNearbyStores = (cep) => __awaiter(void 0, void 0, void 0, function* () {
     try {
         const newData = yield (0, cepData_js_1.cepInfos)(cep.toString());
@@ -109,12 +122,14 @@ const findNearbyStores = (cep) => __awaiter(void 0, void 0, void 0, function* ()
         const userLocation = yield (0, cepToCoordenates_js_1.getCoordenates)(adress);
         const nearbyStores = [];
         const stores = yield getAllStores();
+        //Loop para achar as lojas com menos de 100 km de distancia
         for (let store of stores) {
             store.distancia = (0, distanceCalculate_js_1.harvisineDistanceCalculator)(userLocation.latitude, userLocation.longitude, store.coordenadas.latitude, store.coordenadas.longitude);
             if (store.distancia <= 100) {
                 nearbyStores.push(store);
             }
         }
+        //Ordendando de modo crescente
         nearbyStores.sort((a, b) => {
             if (a.distancia < b.distancia)
                 return -1;
@@ -125,9 +140,10 @@ const findNearbyStores = (cep) => __awaiter(void 0, void 0, void 0, function* ()
     }
     catch (error) {
         loggers_js_1.logger.error("Não foi possivel acessar as lojas proximas: ", error);
-        throw new Error('Someting goes wrong!');
+        throw new Error('Erro no processo de achar as lojas próximas!');
     }
 });
+//Objeto que contém todas as funções
 exports.storeService = {
     createStore,
     createStoreByCep,
